requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

torch==2.1.0
transformers==4.35.0
bitsandbytes==0.41.1
accelerate==0.24.1
sentence-transformers==2.2.2

chromadb==0.4.18

PyPDF2==3.0.1
python-docx==1.1.0

openai-whisper==20231117
pydub==0.25.1
soundfile==0.12.1

Pillow==10.1.0

supabase==2.0.3

python-dotenv==1.0.0
